selenium/requirements.txt
pytest==9.0.2
selenium==4.39.0
webdriver-manager==4.0.2
python-dotenv==1.0.1
